Move app files to utils folder
I didn't like the naming of the folder and the files, so I changed the folder name to utils and put all the exports into the same file

Has the benefit that new developers wont be tempted to look at the unnecessary "routes.js" file, which you never need to change. Better separation of concerns for beginners.